requirements.txt
# Web framework
Flask

# Metadata stripping libraries
piexif         # For JPEG/Exif metadata removal
PyMuPDF        # For PDF parsing and cleaning
python-docx    # For DOCX (Word) file support
nometa         # For XLSX (Excel) support
pyheif
Pillow

# Flask dependencies and helpers
werkzeug       # File handling & request helpers
python-dotenv  # Load environment variables from .env
